feat: move logic to service
