+++ frontend/src/tests/lib/components/accounts/BitcoinAddress.spec.ts
@@ -11,6 +11,8 @@ import {
 } from "$lib/constants/ckbtc-canister-ids.constants";
 import { AppPath } from "$lib/constants/routes.constants";
 import { bitcoinAddressStore } from "$lib/stores/bitcoin.store";
+import { mockMainAccount } from "$tests/mocks/accounts.store.mock";
+import { mockIdentity } from "$tests/mocks/auth.store.mock";
 import {
   mockBTCAddressTestnet,
   mockCkBTCMainAccount,
@@ -39,10 +41,41 @@ describe("BitcoinAddress", () => {
   });
 
   describe("not matching bitcoin address store", () => {
+    let spyGetAddress;
+
     beforeEach(() => {
-      jest
+      spyGetAddress = jest
         .spyOn(minterApi, "getBTCAddress")
         .mockResolvedValue(mockBTCAddressTestnet);
+    });
+
+    it("should load bitcoin address on mount", async () => {
+      render(BitcoinAddress, { props });
+
+      await waitFor(() =>
+        expect(spyGetAddress).toBeCalledWith({
+          identity: mockIdentity,
+          canisterId: CKTESTBTC_MINTER_CANISTER_ID,
+        })
+      );
+    });
+
+    it("should also load bitcoin address on mount if no match", async () => {
+      const data = {
+        identifier: mockMainAccount.identifier,
+        btcAddress: mockBTCAddressTestnet,
+      };
+
+      bitcoinAddressStore.set(data);
+
+      render(BitcoinAddress, { props });
+
+      await waitFor(() =>
+        expect(spyGetAddress).toBeCalledWith({
+          identity: mockIdentity,
+          canisterId: CKTESTBTC_MINTER_CANISTER_ID,
+        })
+      );
     });
 
     it("should render a spinner while loading", async () => {
@@ -60,6 +93,16 @@ describe("BitcoinAddress", () => {
 
     beforeEach(() => {
       bitcoinAddressStore.set(data);
+    });
+
+    it("should not load bitcoin address on mount if already loaded", async () => {
+      const spyGetAddress = jest
+        .spyOn(minterApi, "getBTCAddress")
+        .mockResolvedValue(mockBTCAddressTestnet);
+
+      render(BitcoinAddress, { props });
+
+      await waitFor(() => expect(spyGetAddress).not.toHaveBeenCalled());
     });
 
     it("should not render a spinner when loaded", () => {
